+++ plugins/catalog-backend-module-github/src/providers/GithubEntityProviderConfig.test.ts
@@ -112,6 +112,7 @@ describe('readProviderConfigs', () => {
         },
       },
       schedule: undefined,
+      validateLocationsExist: false,
     });
     expect(providerConfigs[1]).toEqual({
       id: 'providerCustomCatalogPath',
@@ -127,6 +128,7 @@ describe('readProviderConfigs', () => {
         },
       },
       schedule: undefined,
+      validateLocationsExist: false,
     });
     expect(providerConfigs[2]).toEqual({
       id: 'providerWithRepositoryFilter',
@@ -142,6 +144,7 @@ describe('readProviderConfigs', () => {
         },
       },
       schedule: undefined,
+      validateLocationsExist: false,
     });
     expect(providerConfigs[3]).toEqual({
       id: 'providerWithBranchFilter',
@@ -157,6 +160,7 @@ describe('readProviderConfigs', () => {
         },
       },
       schedule: undefined,
+      validateLocationsExist: false,
     });
     expect(providerConfigs[4]).toEqual({
       id: 'providerWithTopicFilter',
@@ -172,6 +176,7 @@ describe('readProviderConfigs', () => {
         },
       },
       schedule: undefined,
+      validateLocationsExist: false,
     });
     expect(providerConfigs[5]).toEqual({
       id: 'providerWithHost',
@@ -186,6 +191,7 @@ describe('readProviderConfigs', () => {
           exclude: undefined,
         },
       },
+      validateLocationsExist: false,
       schedule: undefined,
     });
     expect(providerConfigs[6]).toEqual({
@@ -207,6 +213,38 @@ describe('readProviderConfigs', () => {
           minutes: 3,
         },
       },
+      validateLocationsExist: false,
     });
+  });
+
+  it('defaults validateLocationsExist to false', () => {
+    const config = new ConfigReader({
+      catalog: {
+        providers: {
+          github: {
+            organization: 'test-org',
+          },
+        },
+      },
+    });
+    const providerConfigs = readProviderConfigs(config);
+
+    expect(providerConfigs[0].validateLocationsExist).toEqual(false);
+  });
+
+  it('throws an error when a wildcard catalog path is configured with validation of locations', () => {
+    const config = new ConfigReader({
+      catalog: {
+        providers: {
+          github: {
+            organization: 'test-org',
+            validateLocationsExist: true,
+            catalogPath: '/*/catalog-info.yaml',
+          },
+        },
+      },
+    });
+
+    expect(() => readProviderConfigs(config)).toThrow();
   });
 });