+++ plugins/ilert/src/components/Service/ServiceActionsMenu.tsx
@@ -1,0 +1,68 @@
+/*
+ * Copyright 2021 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { IconButton, Menu, MenuItem, Typography } from '@material-ui/core';
+import MoreVertIcon from '@material-ui/icons/MoreVert';
+import React from 'react';
+import { ilertApiRef } from '../../api';
+import { Service } from '../../types';
+
+import { Link } from '@backstage/core-components';
+import { useApi } from '@backstage/core-plugin-api';
+
+export const ServiceActionsMenu = ({ service }: { service: Service }) => {
+  const ilertApi = useApi(ilertApiRef);
+  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+
+  const handleClick = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorEl(event.currentTarget);
+  };
+
+  const handleCloseMenu = () => {
+    setAnchorEl(null);
+  };
+
+  return (
+    <>
+      <IconButton
+        aria-label="more"
+        aria-controls="long-menu"
+        aria-haspopup="true"
+        onClick={handleClick}
+        size="small"
+      >
+        <MoreVertIcon />
+      </IconButton>
+      <Menu
+        id={`service-actions-menu-${service.id}`}
+        anchorEl={anchorEl}
+        keepMounted
+        open={Boolean(anchorEl)}
+        onClose={handleCloseMenu}
+        PaperProps={{
+          style: { maxHeight: 48 * 5.5 },
+        }}
+      >
+        <MenuItem key="details" onClick={handleCloseMenu}>
+          <Typography variant="inherit" noWrap>
+            <Link to={ilertApi.getServiceDetailsURL(service)}>
+              View in iLert
+            </Link>
+          </Typography>
+        </MenuItem>
+      </Menu>
+    </>
+  );
+};