+++ plugins/xcmetrics/src/components/BuildTimeline/BuildTimeline.test.tsx
@@ -20,14 +20,45 @@ import { BuildTimeline } from './BuildTimeline';
 jest.mock('../../api/XcmetricsClient');
 const client = require('../../api/XcmetricsClient');
 
+jest.mock('recharts', () => {
+  const OriginalModule = jest.requireActual('recharts');
+
+  return {
+    ...OriginalModule,
+    ResponsiveContainer: ({ children }: any) => (
+      <OriginalModule.ResponsiveContainer width={100} aspect={1}>
+        {children}
+      </OriginalModule.ResponsiveContainer>
+    ),
+  };
+});
+
 describe('BuildTimeline', () => {
+  const { ResizeObserver } = window;
+
+  beforeEach(() => {
+    // @ts-expect-error
+    delete window.ResizeObserver;
+    window.ResizeObserver = jest.fn().mockImplementation(() => ({
+      observe: jest.fn(),
+      unobserve: jest.fn(),
+      disconnect: jest.fn(),
+    }));
+  });
+
+  afterEach(() => {
+    window.ResizeObserver = ResizeObserver;
+    jest.restoreAllMocks();
+  });
+
   it('should render', async () => {
     const rendered = await renderInTestApp(
       <BuildTimeline targets={[client.mockTarget]} height={100} width={100} />,
     );
-    expect(
-      await rendered.findByText(client.mockTarget.name),
-    ).toBeInTheDocument();
+
+    const [element] = await rendered.findAllByText(client.mockTarget.name);
+
+    expect(element).toBeInTheDocument();
   });
 
   it('should render a message if no targets are provided', async () => {