+++ plugins/techdocs-module-addons-contrib/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-techdocs-module-addons-contrib
 
+## 1.0.6-next.0
+
+### Patch Changes
+
+- c1784a4980: Replaces in-code uses of `GitHub` with `Github` and deprecates old versions.
+- Updated dependencies
+  - @backstage/core-components@0.12.0-next.0
+  - @backstage/core-plugin-api@1.1.0-next.0
+  - @backstage/integration@1.4.0-next.0
+  - @backstage/integration-react@1.1.6-next.0
+  - @backstage/plugin-techdocs-react@1.0.6-next.0
+  - @backstage/theme@0.2.16
+
 ## 1.0.5
 
 ### Patch Changes